Fix ASN1Parser length parsing
The current length parsing is comparing offset in parsed bytestream to sub-object length.
Sometimes this leads to last branch being taken for an object in middle of ASN1Sequence.
Due to that, excess bytes are included to encodedBytes.

It looks like the intention of the if construct is a bounds check. Fix that.